Added option to print extra network info
print_vtk accepts a third argument of the form unordered_map<string, vector<double>>. The map keys (strings) are the names of the variables to be printed. The vector<double> corresponding to a variable can contain either edge or node values, but must be the same size as either the total node network or the total edge network.